[SEO] Adding more metadata for the rest of the pages
