+++ app/(feed)/following/layout.tsx
@@ -2,7 +2,7 @@ import { Metadata } from 'next';
 import { buildOpenGraphMetadata } from '@/lib/metadata';
 
 export const metadata: Metadata = buildOpenGraphMetadata({
-  title: 'Following Feed',
+  title: 'Following',
   description:
     'Research papers and discussions from scientists and topics you follow on ResearchHub.',
   url: '/following',
