+++ app/(feed)/for-you/layout.tsx
@@ -2,7 +2,7 @@ import { Metadata } from 'next';
 import { buildOpenGraphMetadata } from '@/lib/metadata';
 
 export const metadata: Metadata = buildOpenGraphMetadata({
-  title: 'For You Feed',
+  title: 'For You',
   description:
     'Your personalized feed of scientific research, papers, and discussions based on your interests.',
   url: '/for-you',
