+++ app/(feed)/popular/layout.tsx
@@ -2,7 +2,7 @@ import { Metadata } from 'next';
 import { buildOpenGraphMetadata } from '@/lib/metadata';
 
 export const metadata: Metadata = buildOpenGraphMetadata({
-  title: 'Popular Feed',
+  title: 'Popular',
   description:
     'Discover trending scientific research, papers, and discussions on ResearchHub. See what the community is reading and discussing.',
   url: '/popular',
