+++ app/journal/page.tsx
@@ -5,7 +5,7 @@ import { JournalPage } from '@/components/Journal/JournalPage';
 import { RHJRightSidebar } from '@/components/Journal/RHJRightSidebar';
 
 export const metadata: Metadata = buildOpenGraphMetadata({
-  title: 'ResearchHub Journal',
+  title: 'Journal',
   description:
     'Accelerate science through novel incentive structures that reward authors for reproducible research and compensate peer reviewers for their expertise',
   url: '/journal',
