+++ app/moderators/layout.tsx
@@ -1,55 +1,10 @@
-'use client';
+import { Metadata } from 'next';
+import ModerationClientLayout from './ModerationClientLayout';
 
-import { ReactNode, useEffect } from 'react';
-import { usePathname, useRouter } from 'next/navigation';
-import { LayoutWithRightSidebar } from '../layouts/LayoutWithRightSidebar';
-import { ModerationSidebar } from '@/components/Moderators/ModerationSidebar';
-import { useUser } from '@/contexts/UserContext';
-import { LoadingSkeleton } from '../layouts/components/LoadingSkeleton';
-import { ModerationMenu } from '@/components/Moderators/ModerationSidebar';
+export const metadata: Metadata = {
+  title: 'Moderation',
+};
 
-interface ModerationLayoutProps {
-  readonly children: ReactNode;
-}
-
-const FULL_WIDTH_CLASS =
-  'tablet:!max-w-full content-md:!max-w-full content-lg:!max-w-full content-xl:!max-w-full';
-
-export default function ModerationLayout({ children }: ModerationLayoutProps) {
-  const { user, isLoading } = useUser();
-  const router = useRouter();
-
-  const pathname = usePathname();
-  const isFullWidthTablePage =
-    pathname === '/moderators/referral' || pathname === '/moderators/editors';
-
-  const isModerator = !!user?.isModerator;
-
-  useEffect(() => {
-    if (!isLoading && !isModerator) {
-      router.push('/popular');
-    }
-  }, [isLoading, isModerator, router]);
-
-  if (isLoading) {
-    return (
-      <div className="min-h-screen bg-white">
-        <LoadingSkeleton />
-      </div>
-    );
-  }
-
-  if (!isModerator) {
-    return null;
-  }
-
-  return (
-    <LayoutWithRightSidebar
-      rightSidebar={<ModerationSidebar />}
-      mobileMenu={<ModerationMenu />}
-      className={isFullWidthTablePage ? FULL_WIDTH_CLASS : undefined}
-    >
-      {children}
-    </LayoutWithRightSidebar>
-  );
+export default function ModerationLayout({ children }: { children: React.ReactNode }) {
+  return <ModerationClientLayout>{children}</ModerationClientLayout>;
 }
